change hyphen to dash in page titles
I use the Unicode escape sequence in the JavaScript string just to be careful, since I’m not sure of the extent of browser support of Unicode literals in JavaScript strings.

In the page title in HTML, I use a literal ‘–’, because the page is declared as UTF-8-encoded.

gist.html includes no dash in the title – the title is just the title of the source Gist.