change code tab size from default 8 to 2
Using the CSS3 tab-size property: https://developer.mozilla.org/en-US/docs/CSS/tab-size

This improves the look of code with tabs in it, such as the code in http://bl.ocks.org/4358325.

I chose tab size 2 because that’s the indent size the bl.ocks.org code uses, and because it’s a common tab size for HTML and JavaScript (though size 4 might be more common).

In later work, it might be nice to also set different tab sizes for code blocks in languages with different conventions (this could be done by selecting based on the class of the code element). But I can’t think of any such languages off the top of my head that are likely to appear in code in blocks.