renamed variable `gist` to `gistId`
That’s what the variable represents – not an entire Gist (including its files), just the ID of a Gist.

As far as I know, Gist IDs can only be numbers, so I could call the variable `gistNum`, but I’m calling it `gistId` because Gist that may change that in the future.